move helpers.go into an internal package
